refactor: Have study ids stored in the url of the comparison page (#41)
* refactor[1/2]: Handle view management in the parent StudyView

Rather than having hard links in StudyMenu. This allows us

- Not to reload the data at every change
- Have a dumb component StudyMenu that's easier to understand.

* refactor[2/2]: Don't reload the whole app every time we change the url

Otherwise the comparaison page reloads whenever we change the query.
I don't know why this has been put here in the first place, I don't think it's good practice and I don't see a real counterargument in removing it. Let me know if I'm missing something.

Note: Since the study menu relied on the router component `StudyView` to be reloaded whenever the page url changed the slightest, this wouldn't have been possible with the previous refactor.

* boyscout [1/2]: Remove useless isLocal

* boyscout [2/2] clarify currency flow

Base ourselves on the router.query, as for the view

* feat [1/2] Use query array to pass all ids of studies we want to compare

* feat [2/2] Load studies in the url dynamically

instead of specific country/product studies

* fix: Use string util to encode id list

Otherwise the type would not be an array if there was only 1 study id in the url

* feat: Display "No study is selected" when there's no study on the comaprison page

Note: We'll invite the user to add some on the next PR

Bonus: I'm adding a min-height and tweaking the flew to the app so that the footer always is at the bottom even when the page is too small

* fixup boyscout2

* fix: Don't display empty studies when we are loading