fix: use `doSearch` to sort rows
Because the slot-mode we use for checkboxes insertion do not support sorting correctly.
So we manually sort the rows in the component on `@doSearch`

Note: We should inject directly the items with the formatted field to the `DataTable` component as well so the sorting can rely on the `field` keys of the columns prop.
This is not ideal as one user of `dataTable` _must_ have this prop, but this works well for the 2 project usecases  here